% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fxns.R
\name{euclidean_dist}
\alias{euclidean_dist}
\title{Euclidean Distance}
\usage{
euclidean_dist(a, b)
}
\arguments{
\item{a}{A vector}

\item{b}{A vector}
}
\value{
A value representing the distance
}
\description{
This function calculates the Euclidean distance between two points
}
