% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds_plot.R
\name{fds_plot}
\alias{fds_plot}
\title{FDS Plot}
\usage{
fds_plot(design1, design2, y_lim = NULL, col_v = col_v, title = "")
}
\arguments{
\item{design1}{an experimental design for comparison.}

\item{design2}{an experimental design for comparison.}

\item{y_lim}{ignore.}

\item{col_v}{a vector of strings containing a number of colors equal to the
number of designs being plotted.}

\item{title}{a string containing the desired title for the FDS plot.}
}
\value{
an FDS plot.
}
\description{
fds_plot takes two experiment designs and creates a Fraction of Design Space
visualization that compares relative prediction variance across the design
space. A lower prediction variance is considered a a desirable quality in an
experiment.
}
