% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds_plot.R
\name{make_model_mat}
\alias{make_model_mat}
\title{make_model_mat Function}
\usage{
make_model_mat(x)
}
\arguments{
\item{x}{an experiment design in the form of a matrix.}
}
\value{
a model matrix of the input design.
}
\description{
make_model_mat takes an experiment's design matrix and extends it into the
model matrix. The function creates a second order model which includes main
effects, interaction, and squared terms.
}
