% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds_plot.R
\name{make_model_vec}
\alias{make_model_vec}
\title{make_model_vec Function}
\usage{
make_model_vec(x)
}
\arguments{
\item{x}{a vector or matrix object that represents an experiment design}
}
\value{
a single vector
}
\description{
makeModelVec generates a single vector of the design matrix for computation
purposes in makePredVec.
}
