% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds_plot.R
\name{make_pred_vec}
\alias{make_pred_vec}
\title{make_pred_vec Function}
\usage{
make_pred_vec(x)
}
\arguments{
\item{x}{candidate design being compared.}
}
\value{
a list of ordered prediction variance of the randomized designs.
}
\description{
make_pred_vec creates a list of prediction variances of randomized designs
for the fds_plot function.
}
