% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds_plot.R
\name{make_random_grid}
\alias{make_random_grid}
\title{make_random_grid generates random designs}
\usage{
make_random_grid(dim = 1)
}
\arguments{
\item{dim}{the dimension of the grid. Accepts only values of 1 or 2.}
}
\value{
a randomized design with user specified number of parameters.
}
\description{
make_random_grid generates random designs
}
