% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds_plot.R
\name{makeModelVec}
\alias{makeModelVec}
\title{makeModelVec generates a single vector of the design matrix}
\usage{
makeModelVec(X)
}
\arguments{
\item{X}{a vector or matrix object that represents an experiment design}
}
\value{
something
}
\description{
makeModelVec generates a single vector of the design matrix
}
