% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds_plot.R
\name{makePredVec}
\alias{makePredVec}
\title{makePredVec creates a list of prediction variances of randomized designs for
the FDS plot}
\usage{
makePredVec(X)
}
\arguments{
\item{X}{candidate design being compared}
}
\value{
a list of ordered prediction variance of the randomized designs
}
\description{
makePredVec creates a list of prediction variances of randomized designs for
the FDS plot
}
