% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fxns.R
\name{min_dist}
\alias{min_dist}
\title{Minimum Distance Function}
\usage{
min_dist(X, dist_fx)
}
\arguments{
\item{X}{The matrix for which the minimum distance between all rows will
be found}

\item{dist_fx}{The distance between two rows of a matrix (check this)}
}
\value{
The minimum distance between all rows in a matrix
}
\description{
This function determines the minimum distance between all rows
in a matrix
}
