% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design.R
\name{plot_design}
\alias{plot_design}
\title{Plot Design Function}
\usage{
plot_design(data_row)
}
\arguments{
\item{data_row}{A row of data.}
}
\value{
A plot of the specified geometry.
}
\description{
This function plots hypercube and simplex geometries.
}
