% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design.R
\name{plot_hypercube}
\alias{plot_hypercube}
\title{Plot Hypercube Function}
\usage{
plot_hypercube(data_row)
}
\arguments{
\item{data_row}{A row of data.}
}
\value{
A plot of the hypercube.
}
\description{
This function plots the specified distances on a hypercube geometry.
}
