% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_design.R
\name{plot_simplex}
\alias{plot_simplex}
\title{Plot Simplex Function}
\usage{
plot_simplex(data_row)
}
\arguments{
\item{data_row}{A row of data.}
}
\value{
A plot of the simplex.
}
\description{
This function plots the specified distances on a simplex geometry
}
