% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_fxns.R
\name{print_design}
\alias{print_design}
\title{print_design}
\usage{
print_design(design_row)
}
\arguments{
\item{design_row}{a row in the DoEAna data set}
}
\value{
a matrix of the design.
}
\description{
print_design
}
