% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_efficiency.R
\name{relative_efficiency}
\alias{relative_efficiency}
\title{Relative Efficiency}
\usage{
relative_efficiency(m_row, e_row, c_row)
}
\arguments{
\item{m_row}{Generated with mahattan distance}

\item{e_row}{Generated with euclidean distance}

\item{c_row}{Generated with chebyshev distance}
}
\value{
A 3x3 data frame efficiency table
}
\description{
This function takes in the different generated distances
and selects specific traits from each type.  With the
selected values, relative efficiency is done for each
distance on each score "(M(x), E(x), C(x)).  The output
will be a table with all efficiencies.
}
