% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_design.R
\name{select_design}
\alias{select_design}
\title{Select Design Function}
\usage{
select_design(d_data, geometry, distance, n)
}
\arguments{
\item{d_data}{data.frame that will be subsetted to a certian row.}

\item{geometry}{Geometry of the design space, either 'simplex' or
'hypercube'}

\item{distance}{Distance metric used in design, either 'manhattan',
'chebyshev', or 'euclidean'}

\item{n}{Number of design points, either 10, 20, or 30}
}
\value{
Specified row of the DoEAna dataframe that can be used for further
analysis
}
\description{
Select Design Function
}
