% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_fxns.R
\name{chebyshev_dist}
\alias{chebyshev_dist}
\title{Chebyshev function}
\usage{
chebyshev_dist(A, B)
}
\arguments{
\item{A}{A vector}

\item{B}{A vector}
}
\value{
A value which is the distance
}
\description{
This function calculates the Chebyshev distance between two points
}
