% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds_plot.R
\name{fds_plot}
\alias{fds_plot}
\title{Title}
\usage{
fds_plot(design1, design2, Ylim = NULL, colV = colv, Main = "")
}
\arguments{
\item{Ylim}{ignore}

\item{colV}{a vector of strings containing a number of colors equal to the
number of designs being plotted}

\item{Main}{a string containing the desired title for the FDS plot}

\item{design}{an experimental design}
}
\value{
an FDS plot
}
\description{
Title
}
