% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds_plot.R
\name{makeModelMat}
\alias{makeModelMat}
\title{makeModelMat takes an experiment's design matrix and extends it into the
model matrix. The function creates a second order model which includes main
effects, interaction, and squared terms.}
\usage{
makeModelMat(X)
}
\arguments{
\item{X}{yep}
}
\value{
a model matrix of the input design.

model matrix
}
\description{
makeModelMat takes an experiment's design matrix and extends it into the
model matrix. The function creates a second order model which includes main
effects, interaction, and squared terms.
}
